# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookify",
        "description": "O objetivo do projeto é fornecer uma API RESTful para permitir que os usuários realizem diversas operações relacionadas a livros",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/{id?}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Listar todos os Livros",
                "operationId": "listBooks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro (opcional)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação dos resultados",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de itens por página para paginação dos resultados",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenação dos resultados",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filtro para buscar por título, autor, etc.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de solicitação inválida"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            }
        },
        "/api/books": {
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Criar Livro",
                "operationId": "createBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "genres_id": {
                                        "type": "integer"
                                    },
                                    "edition": {
                                        "type": "integer"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "pages": {
                                        "type": "integer"
                                    },
                                    "format": {
                                        "type": "string"
                                    },
                                    "license": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de solicitação inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualizar Livro",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "genres_id": {
                                        "type": "integer"
                                    },
                                    "edition": {
                                        "type": "integer"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "pages": {
                                        "type": "integer"
                                    },
                                    "format": {
                                        "type": "string"
                                    },
                                    "license": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de solicitação inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Excluir Livro",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "required": [
                    "id",
                    "title",
                    "author_id",
                    "genres_id",
                    "edition",
                    "year",
                    "pages",
                    "format",
                    "license"
                ],
                "properties": {
                    "id": {
                        "description": "ID do livro",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Título do livro",
                        "type": "string"
                    },
                    "author_id": {
                        "description": "ID do autor do livro",
                        "type": "integer",
                        "format": "int64"
                    },
                    "genres_id": {
                        "description": "ID do gênero do livro",
                        "type": "integer",
                        "format": "int64"
                    },
                    "edition": {
                        "description": "Edição do livro",
                        "type": "integer"
                    },
                    "year": {
                        "description": "Ano de publicação do livro",
                        "type": "integer"
                    },
                    "pages": {
                        "description": "Número de páginas do livro",
                        "type": "integer"
                    },
                    "format": {
                        "description": "Formato do livro",
                        "type": "string"
                    },
                    "license": {
                        "description": "Licença do livro",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do livro",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação do livro",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização do livro",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}